+++ CHANGES.rst
@@ -37,8 +37,10 @@ Release history
 - ``Model`` now has a ``connection_decode_neurons`` attribute that maps ``Connection``
   objects that require decode neurons to the corresponding ``Ensemble`` objects
   implementing them. (`#309`_)
-- Added the ``GreedyComms`` allocator, which reduces inter-chip communication, speeding
-  up networks with high traffic between chips. (`#309`_)
+- Added the ``GreedyComms`` and ``PartitionComms`` allocators, which reduce inter-chip
+  communication, speeding up networks with high traffic between chips.
+  ``PartitionComms`` typically finds a more optimal partitioning than ``GreedyComms``,
+  but does require the ``nxmetis`` package. (`#309`_)
 
 **Changed**